Move ParentsModel into parents subdirectory.
