+++ client/components/edit/EditHome.test.tsx
@@ -13,7 +13,7 @@ jest.mock("../../context/EditorContext", () => ({
 
 jest.mock("./children/ChildList", () => () => "ChildList");
 jest.mock("./EditorSnackbar", () => () => "EditorSnackbar");
-jest.mock("./ParentsModal", () => () => "ParentsModal");
+jest.mock("./parents/ParentsModal", () => () => "ParentsModal");
 jest.mock("./StateModal", () => () => "StateModal");
 
 describe("EditHome", () => {
