+++ client/components/edit/ObjectEditor.tsx
@@ -7,7 +7,7 @@ import Grid from "@mui/material/Grid";
 import Link from "next/link";
 import DatastreamList from "./datastream/DatastreamList";
 import DatastreamModal from "./datastream/DatastreamModal";
-import ParentsModal from "./ParentsModal";
+import ParentsModal from "./parents/ParentsModal";
 import StateModal from "./StateModal";
 import { useEditorContext } from "../../context/EditorContext";
 import EditorSnackbar from "./EditorSnackbar";
